+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcHdDsp/DmcHdDspAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcHdDsp
+{
+	[KrangSettings("DmcHdDsp", typeof(DmcHdDspAdapter))]
+	public sealed class DmcHdDspAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}