+++ ICD.Connect.Routing/Pathfinding/DefaultPathFinder.cs
@@ -2,10 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using ICD.Common.Utils;
-using ICD.Common.Utils.Collections;
 using ICD.Common.Utils.Extensions;
-using ICD.Common.Utils.Services;
-using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.Routing.Connections;
 using ICD.Connect.Routing.Endpoints;
 using ICD.Connect.Routing.RoutingGraphs;
@@ -14,16 +11,8 @@ namespace ICD.Connect.Routing.PathFinding
 {
 	public sealed class DefaultPathFinder : AbstractPathFinder
 	{
-		private readonly IRoutingGraph m_RoutingGraph;
 		private readonly IConnectionsCollection m_Connections;
 		private readonly int m_RoomId;
-
-		private ILoggerService m_CachedLogger;
-
-		public ILoggerService Logger
-		{
-			get { return m_CachedLogger = m_CachedLogger ?? ServiceProvider.TryGetService<ILoggerService>(); }
-		}
 
 		/// <summary>
 		/// Constructor.
@@ -33,10 +22,11 @@ namespace ICD.Connect.Routing.PathFinding
 			if (routingGraph == null)
 				throw new ArgumentNullException("routingGraph");
 
-			m_RoutingGraph = routingGraph;
-			m_Connections = m_RoutingGraph.Connections;
+			m_Connections = routingGraph.Connections;
 			m_RoomId = roomId;
 		}
+
+		#region Methods 
 
 		/// <summary>
 		/// Returns the best paths for the given builder queries.
@@ -49,6 +39,91 @@ namespace ICD.Connect.Routing.PathFinding
 				throw new ArgumentNullException("queries");
 
 			return queries.SelectMany(q => FindPaths(q));
+		}
+
+		/// <summary>
+		/// Returns true if there is a valid path for all of the defined queries.
+		/// </summary>
+		/// <param name="queries"></param>
+		/// <returns></returns>
+		public override bool HasPaths(IEnumerable<PathBuilderQuery> queries)
+		{
+			if (queries == null)
+				throw new ArgumentNullException("queries");
+
+			return queries.All(q => HasPaths(q));
+		}
+
+		#endregion
+
+		#region Private Methods
+
+		/// <summary>
+		/// Returns true if there is a valid path for the given query.
+		/// </summary>
+		/// <param name="query"></param>
+		/// <returns></returns>
+		private bool HasPaths(PathBuilderQuery query)
+		{
+			if (query == null)
+				throw new ArgumentNullException("query");
+
+			EndpointInfo[] source = query.GetStart().ToArray();
+			EndpointInfo[][] destinations = query.GetEnds().ToArray();
+
+			foreach (eConnectionType flag in EnumUtils.GetFlagsExceptNone(query.Type))
+			{
+				if (!HasPaths(source, destinations, flag))
+					return false;
+			}
+
+			return true;
+		}
+
+		/// <summary>
+		/// Returns true if there is a valid path from the source to each destination.
+		/// </summary>
+		/// <param name="source"></param>
+		/// <param name="destinations"></param>
+		/// <param name="flag"></param>
+		/// <returns></returns>
+		private bool HasPaths(EndpointInfo[] source, EndpointInfo[][] destinations, eConnectionType flag)
+		{
+			if (source == null)
+				throw new ArgumentNullException("source");
+
+			if (destinations == null)
+				throw new ArgumentNullException("destinations");
+
+			if (!EnumUtils.HasSingleFlag(flag))
+				throw new ArgumentException("Connection type has multiple flags", "flag");
+
+			foreach (EndpointInfo[] destination in destinations)
+			{
+				bool destinationHasPath = false;
+
+				foreach (EndpointInfo destinationEndpoint in destination)
+				{
+					foreach (EndpointInfo sourceEndpoint in source)
+					{
+						Connection outputConnection =
+							m_Connections.GetOutputConnection(sourceEndpoint, destinationEndpoint, flag);
+						if (outputConnection == null)
+							continue;
+
+						destinationHasPath = true;
+						break;
+					}
+
+					if (destinationHasPath)
+						break;
+				}
+
+				if (!destinationHasPath)
+					return false;
+			}
+
+			return true;
 		}
 
 		/// <summary>
@@ -90,8 +165,6 @@ namespace ICD.Connect.Routing.PathFinding
 			if (!EnumUtils.HasSingleFlag(flag))
 				throw new ArgumentException("Connection type has multiple flags", "flag");
 
-			IcdHashSet<EndpointInfo[]> notFound = new IcdHashSet<EndpointInfo[]>(destinations);
-
 			// Get the output connections for the source
 			Connection[] sourceConnections =
 				source.Select(e => m_Connections.GetOutputConnection(e, flag))
@@ -112,23 +185,9 @@ namespace ICD.Connect.Routing.PathFinding
 					if (path == null)
 						continue;
 
-					notFound.Remove(destination);
-
 					yield return path;
 					break;
 				}
-			}
-
-			// Log errors
-			foreach (EndpointInfo[] destination in notFound)
-			{
-				string sourceText = EndpointInfo.ArrayRangeFormat(source);
-				string destinationText = EndpointInfo.ArrayRangeFormat(destination);
-
-				string message = string.Format("{0} failed to find {1} path from {2} to {3}", GetType().Name, flag, sourceText,
-											   destinationText);
-
-				Logger.AddEntry(eSeverity.Error, message);
 			}
 		}
 
@@ -161,14 +220,14 @@ namespace ICD.Connect.Routing.PathFinding
 		/// </summary>
 		/// <param name="source"></param>
 		/// <param name="finalDestinations"></param>
-		/// <param name="inputConnection"></param>
+		/// <param name="sourceOutputConnection"></param>
 		/// <param name="flag"></param>
 		/// <returns></returns>
 		private IEnumerable<Connection> GetConnectionChildren(EndpointInfo source, IEnumerable<EndpointInfo> finalDestinations,
-															  Connection inputConnection, eConnectionType flag)
+															  Connection sourceOutputConnection, eConnectionType flag)
 		{
-			if (inputConnection == null)
-				throw new ArgumentNullException("inputConnection");
+			if (sourceOutputConnection == null)
+				throw new ArgumentNullException("sourceOutputConnection");
 
 			if (finalDestinations == null)
 				throw new ArgumentNullException("finalDestinations");
@@ -177,12 +236,14 @@ namespace ICD.Connect.Routing.PathFinding
 				throw new ArgumentException("ConnectionType has multiple flags", "flag");
 
 			return
-				m_Connections.GetOutputConnections(inputConnection.Destination.GetDeviceControlInfo(),
+				m_Connections.GetOutputConnections(sourceOutputConnection.Destination.GetDeviceControlInfo(),
 				                                   finalDestinations,
 				                                   flag)
 				             .Where(c =>
 				                    c.IsAvailableToSourceDevice(source.Device) &&
 				                    c.IsAvailableToRoom(m_RoomId));
 		}
+
+		#endregion
 	}
 }